todolist_DRF/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-environ==0.12.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
sqlparse==0.5.3
typing_extensions==4.14.0
